requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2
django-cors-headers==3.7.0
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
pdf2image==1.14.0
Pillow==8.2.0
psycopg2==2.8.6
pytube==10.7.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
whitenoise==5.2.0
